requirements.txt
torch~=1.10.0
numpy~=1.21.4
matplotlib~=3.5.0
Pillow~=8.4.0
torchvision~=0.11.1
opencv-python~=4.5.4.60
moviepy~=1.0.3
tqdm~=4.62.3